Bug 1688120 - Remove UB due to unnecessary integer conversion r=peterv
Differential Revision: https://phabricator.services.mozilla.com/D106577

UltraBlame original commit: 77247863a2ea3156aaab18dec3565d82513ac09f

+++ js/xpconnect/src/XPCWrappedNativeInfo.cpp
@@ -338,7 +338,7 @@ already_AddRefed<XPCNativeInterface> XPCNativeInterface::NewInstance(
 
   
   
-  int size = sizeof(XPCNativeInterface);
+  size_t size = sizeof(XPCNativeInterface);
   if (realTotalCount > 1) {
     size += (realTotalCount - 1) * sizeof(XPCNativeMember);
   }
